fix: ssr h1 on /layer-2/networks for seo
