+++ app/[locale]/layer-2/networks/page.tsx
@@ -8,7 +8,12 @@ import {
 
 import type { Lang, PageParams } from "@/lib/types"
 
+import CalloutSSR from "@/components/CalloutSSR"
+import { ContentHero, type ContentHeroProps } from "@/components/Hero"
 import I18nProvider from "@/components/I18nProvider"
+import MainArticle from "@/components/MainArticle"
+import NetworkMaturity from "@/components/NetworkMaturity"
+import { ButtonLink } from "@/components/ui/buttons/Button"
 
 import { getAppPageContributorInfo } from "@/lib/utils/contributors"
 import { getMetadata } from "@/lib/utils/metadata"
@@ -28,6 +33,9 @@ import {
   getGrowThePieMasterData,
   getL2beatData,
 } from "@/lib/data"
+import heroImg from "@/public/images/heroes/layer-2-hub-hero.png"
+import Callout2Image from "@/public/images/layer-2/layer-2-walking.png"
+import Callout1Image from "@/public/images/man-and-dog-playing.png"
 
 const Page = async (props: { params: Promise<PageParams> }) => {
   const params = await props.params
@@ -137,6 +145,16 @@ const Page = async (props: { params: Promise<PageParams> }) => {
     locale as Lang
   )
 
+  const t = await getTranslations("page-layer-2-networks")
+  const tCommon = await getTranslations("common")
+
+  const heroProps: ContentHeroProps = {
+    breadcrumbs: { slug: "/layer-2/networks", startDepth: 1 },
+    heroImg,
+    title: tCommon("nav-networks-explore-networks-label"),
+    description: t("page-layer-2-networks-hero-description"),
+  }
+
   return (
     <I18nProvider locale={locale} messages={messages}>
       <Layer2NetworksPageJsonLD
@@ -144,9 +162,80 @@ const Page = async (props: { params: Promise<PageParams> }) => {
         layer2Data={layer2DataCompiled}
         contributors={contributors}
       />
-      <Suspense>
-        <Layer2Networks {...layer2NetworksProps} />
-      </Suspense>
+      <MainArticle className="relative flex flex-col">
+        <ContentHero {...heroProps} />
+
+        <Suspense>
+          <Layer2Networks {...layer2NetworksProps} />
+        </Suspense>
+
+        <div id="more-advanced-cta" className="w-full px-8 py-9">
+          <div className="flex flex-col gap-8 bg-main-gradient px-12 py-14">
+            <h3>{t("page-layer-2-networks-more-advanced-title")}</h3>
+            <div className="flex max-w-[768px] flex-col gap-8">
+              <p>
+                {t("page-layer-2-networks-more-advanced-descripton-1")}{" "}
+                <strong>
+                  {t("page-layer-2-networks-more-advanced-descripton-2")}
+                </strong>
+              </p>
+              <p>{t("page-layer-2-networks-more-advanced-descripton-3")}</p>
+            </div>
+            <div className="flex flex-col gap-6 sm:flex-row">
+              <ButtonLink href="https://l2beat.com">
+                {t("page-layer-2-networks-more-advanced-link-1")}
+              </ButtonLink>
+              <ButtonLink href="https://growthepie.com">
+                {t("page-layer-2-networks-more-advanced-link-2")}
+              </ButtonLink>
+            </div>
+          </div>
+        </div>
+
+        <NetworkMaturity />
+
+        <div
+          id="callout-cards"
+          className="flex w-full flex-col px-8 py-9 lg:flex-row lg:gap-16"
+        >
+          <CalloutSSR
+            image={Callout1Image}
+            title={t("page-layer-2-networks-callout-1-title")}
+            description={t("page-layer-2-networks-callout-1-description")}
+          >
+            <div>
+              <ButtonLink
+                href="/layer-2/"
+                customEventOptions={{
+                  eventCategory: "l2_networks",
+                  eventAction: "button_click",
+                  eventName: "bottom_hub",
+                }}
+              >
+                {tCommon("learn-more")}
+              </ButtonLink>
+            </div>
+          </CalloutSSR>
+          <CalloutSSR
+            image={Callout2Image}
+            title={t("page-layer-2-networks-callout-2-title")}
+            description={t("page-layer-2-networks-callout-2-description")}
+          >
+            <div>
+              <ButtonLink
+                href="/layer-2/learn/"
+                customEventOptions={{
+                  eventCategory: "l2_networks",
+                  eventAction: "button_click",
+                  eventName: "bottom_learn",
+                }}
+              >
+                {tCommon("learn-more")}
+              </ButtonLink>
+            </div>
+          </CalloutSSR>
+        </div>
+      </MainArticle>
     </I18nProvider>
   )
 }
